+++ gazebo/gui/building/BuildingModelManip.cc
@@ -18,6 +18,7 @@
 #include "gazebo/rendering/Visual.hh"
 #include "gazebo/common/Exception.hh"
 #include "gazebo/math/Quaternion.hh"
+#include "gazebo/gui/building/BuildingEditorEvents.hh"
 #include "gazebo/gui/building/BuildingMaker.hh"
 #include "gazebo/gui/building/BuildingModelManip.hh"
 
@@ -28,6 +29,11 @@ using namespace gui;
 BuildingModelManip::BuildingModelManip()
 {
   this->parent = NULL;
+  this->level = 0;
+
+  this->connections.push_back(
+  gui::editor::Events::ConnectChangeBuildingLevel(
+    boost::bind(&BuildingModelManip::OnChangeLevel, this, _1)));
 }
 
 /////////////////////////////////////////////////
@@ -64,6 +70,12 @@ rendering::VisualPtr BuildingModelManip::GetVisual() const
 common::Color BuildingModelManip::GetColor() const
 {
   return this->color;
+}
+
+/////////////////////////////////////////////////
+std::string BuildingModelManip::GetTexture() const
+{
+  return this->texture;
 }
 
 /////////////////////////////////////////////////
@@ -269,9 +281,21 @@ void BuildingModelManip::OnRotationChanged(double _roll, double _pitch,
 }
 
 /////////////////////////////////////////////////
+void BuildingModelManip::OnLevelChanged(int _level)
+{
+  this->SetLevel(_level);
+}
+
+/////////////////////////////////////////////////
 void BuildingModelManip::OnColorChanged(QColor _color)
 {
   this->SetColor(_color);
+}
+
+/////////////////////////////////////////////////
+void BuildingModelManip::OnTextureChanged(QString _texture)
+{
+  this->SetTexture(_texture);
 }
 
 /////////////////////////////////////////////////
@@ -342,7 +366,44 @@ void BuildingModelManip::SetColor(QColor _color)
 }
 
 /////////////////////////////////////////////////
+void BuildingModelManip::SetTexture(QString _texture)
+{
+  // TODO For now setting existing material scripts.
+  // Add support for custom textures.
+  this->texture = "Gazebo/Grey";
+  if (_texture == ":wood.jpg")
+    this->texture = "Gazebo/Wood";
+  else if (_texture == ":tiles.jpg")
+    this->texture = "Gazebo/CeilingTiled";
+
+  this->visual->GetParent()->SetMaterial(this->texture);
+}
+
+/////////////////////////////////////////////////
 void BuildingModelManip::SetTransparency(float _transparency)
 {
   this->visual->GetParent()->SetTransparency(_transparency);
+}
+
+/////////////////////////////////////////////////
+void BuildingModelManip::SetLevel(const int _level)
+{
+  this->level = _level;
+}
+
+/////////////////////////////////////////////////
+int BuildingModelManip::GetLevel() const
+{
+  return this->level;
+}
+
+/////////////////////////////////////////////////
+void BuildingModelManip::OnChangeLevel(int _level)
+{
+  if (this->level > _level)
+    this->SetTransparency(1.0);
+  else if (this->level < _level)
+    this->SetTransparency(0.0);
+  else
+    this->SetTransparency(0.4);
 }