+++ gazebo/gui/building/EditorView.cc
@@ -273,19 +273,19 @@ void EditorView::mouseReleaseEvent(QMouseEvent *_event)
       }
       else if (this->drawMode == STAIRS)
       {
-        this->stairsList.push_back(dynamic_cast<StairsItem *>(
-            this->currentMouseItem));
+        StairsItem *stairsItem = dynamic_cast<StairsItem *>(
+            this->currentMouseItem);
+        stairsItem->Set3dTexture(QString(""));
+        stairsItem->Set3dColor(Qt::white);
+        this->stairsList.push_back(stairsItem);
         if ((this->currentLevel) < static_cast<int>(floorList.size()))
         {
-          EditorItem *item = dynamic_cast<EditorItem *>(this->currentMouseItem);
-          this->buildingMaker->AttachManip(this->itemToVisualMap[item],
+          this->buildingMaker->AttachManip(this->itemToVisualMap[stairsItem],
               this->itemToVisualMap[floorList[this->currentLevel]]);
         }
       }
-
-      // Select -> deselect to trigger change
-      this->currentMouseItem->setSelected(true);
-      this->currentMouseItem->setSelected(false);
+      dynamic_cast<EditorItem *>(this->currentMouseItem)->
+          SetHighlighted(false);
 
       this->drawMode = NONE;
       this->drawInProgress = false;
@@ -567,6 +567,8 @@ void EditorView::DeleteItem(EditorItem *_item)
   if (!_item)
     return;
 
+  this->buildingMaker->DetachAllChildren(this->itemToVisualMap[_item]);
+
   if (_item->GetType() == "WallSegment")
   {
     WallSegmentItem *wallSegmentItem = dynamic_cast<WallSegmentItem *>(_item);
@@ -679,9 +681,9 @@ void EditorView::DrawWall(const QPoint &_pos)
     this->snapGrabberCurrent = NULL;
 
     wallSegmentItem = dynamic_cast<WallSegmentItem*>(this->currentMouseItem);
-    // Select -> deselect to trigger change
-    wallSegmentItem->setSelected(true);
-    wallSegmentItem->setSelected(false);
+    wallSegmentItem->Set3dTexture(QString(""));
+    wallSegmentItem->Set3dColor(Qt::white);
+    wallSegmentItem->SetHighlighted(false);
     wallSegmentList.push_back(wallSegmentItem);
     if (wallSegmentItem->GetLevel() > 0)
     {
@@ -1019,10 +1021,9 @@ void EditorView::OnAddLevel()
     this->itemToVisualMap[wallSegmentItem] = wallSegmentName;
 
     floorItem->AttachWallSegment(wallSegmentItem);
-
-    // Select -> deselect to trigger change
-    wallSegmentItem->setSelected(true);
-    wallSegmentItem->setSelected(false);
+    wallSegmentItem->Set3dTexture(QString(""));
+    wallSegmentItem->Set3dColor(Qt::white);
+    wallSegmentItem->SetHighlighted(false);
   }
 
   // Clone linked grabber relations
@@ -1066,6 +1067,9 @@ void EditorView::OnAddLevel()
   this->itemToVisualMap[floorItem] = floorName;
   this->scene()->addItem(floorItem);
   this->floorList.push_back(floorItem);
+  floorItem->Set3dTexture(QString(""));
+  floorItem->Set3dColor(Qt::white);
+  floorItem->SetHighlighted(false);
 }
 
 /////////////////////////////////////////////////
@@ -1081,6 +1085,7 @@ void EditorView::DeleteLevel(int _level)
       || _level >= static_cast<int>(this->levels.size()))
     return;
 
+  // Delete current level and move to level below or above
   int newLevelIndex = _level - 1;
   if (newLevelIndex < 0)
     newLevelIndex = _level + 1;
@@ -1212,6 +1217,7 @@ void EditorView::OnOpenLevelInspector()
   {
     this->levelInspector->floorWidget->show();
     this->levelInspector->SetFloorColor(floorItem->Get3dColor());
+    this->levelInspector->SetFloorTexture(floorItem->Get3dTexture());
   }
   else
   {
@@ -1229,9 +1235,14 @@ void EditorView::OnLevelApply()
 
   std::string newLevelName = dialog->GetLevelName();
   this->levels[this->currentLevel]->name = newLevelName;
-  this->levels[this->currentLevel]->floorItem->Set3dColor(dialog->
-      GetFloorColor());
-  this->levels[this->currentLevel]->floorItem->FloorChanged();
+  FloorItem *floorItem = this->levels[this->currentLevel]->floorItem;
+  if (floorItem)
+  {
+    floorItem->Set3dTexture(dialog->GetFloorTexture());
+    floorItem->Set3dColor(dialog->GetFloorColor());
+    floorItem->Set3dTransparency(0.4);
+    floorItem->FloorChanged();
+  }
   gui::editor::Events::updateLevelWidget(this->currentLevel, newLevelName);
 }