+++ gazebo/gui/building/StairsItem.cc
@@ -258,7 +258,8 @@ void StairsItem::OnApply()
     this->StepsChanged();
   }
   // this->stairsElevation = dialog->GetElevation();
-  this->visual3dColor = dialog->GetColor();
+  this->Set3dTexture(dialog->GetTexture());
+  this->Set3dColor(dialog->GetColor());
   this->StairsChanged();
 }
 
@@ -275,6 +276,7 @@ void StairsItem::OnOpenInspector()
   startPos.setY(-startPos.y());
   this->inspector->SetStartPosition(startPos);
   this->inspector->SetColor(this->visual3dColor);
+  this->inspector->SetTexture(this->visual3dTexture);
   this->inspector->move(QCursor::pos());
   this->inspector->show();
 }
@@ -293,7 +295,6 @@ void StairsItem::StairsChanged()
   emit HeightChanged(this->stairsHeight);
   emit PositionChanged(this->stairsPos.x(), this->stairsPos.y(),
       this->levelBaseHeight + this->stairsElevation);
-  emit ColorChanged(this->visual3dColor);
 }
 
 /////////////////////////////////////////////////